move benchmark main file to benchmarking folder, add elements to gitignore
